Add options to configure external-dns
"ExternalDNS synchronizes exposed Kubernetes Services and Ingresses with
DNS providers."[1]

We need this to ease access to development clusters, by synchronizing
Services and Ingresses into a internal DNS zone.

This commit adds the final wiring, so the zone can be set as part of the
cluster provisioning. See [2] and [3], for the chart and yggdrasil
changes.

[1] https://github.com/kubernetes-sigs/external-dns
[2] https://github.com/distributed-technologies/helm-charts/pull/264
[3] https://github.com/distributed-technologies/yggdrasil/pull/75