Fix broken FLUX_GIT_BRANCH option on initial run
The option was never used if the git repository didn't already exsits
and if the repository did exists, after fetching the branch was just
reset to the current revision (missing origin/).

While we are it, only the specified branch is cloned, as this entails
other branches aren't fetched, the branch is now part of the cache key,
to support switching branches if the git repository already exists.

Fixes: 55174cc ("Add support for the Flux GitOps operator")